Code cleanup of fake driver v2
