+++ Tests/TerminalGuiFluentTesting/FakeDriver/FakeApplicationLifecycle.cs
@@ -1,0 +1,16 @@
+﻿#nullable enable
+namespace Terminal.Gui.Drivers;
+
+#pragma warning disable CS1591
+internal class FakeApplicationLifecycle (IApplication origApp, CancellationTokenSource hardStop) : IDisposable
+{
+    /// <inheritdoc/>
+    public void Dispose ()
+    {
+        hardStop.Cancel ();
+
+        Application.Top?.Dispose ();
+        Application.Shutdown ();
+        ApplicationImpl.ChangeInstance (origApp);
+    }
+}
