+++ Tests/TerminalGuiFluentTesting/FakeDriver/FakeNetInput.cs
